Add dedicated Microsoft SSO login page
