+++ README.md
@@ -43,3 +43,10 @@ Preview the production build locally:
 ```bash
 npm run preview
 ```
+
+## Single Sign-On
+
+This project now includes a Microsoft single sign-on flow. When the app starts,
+unauthenticated users see a dedicated login page with a "Sign in with Microsoft"
+button. Update `src/authConfig.ts` with your Azure application client ID before
+running the app.
