+++ package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "lucide-react": "^0.344.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "@azure/msal-browser": "^3.7.0",
+    "@azure/msal-react": "^2.0.20"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.1",
